Split benchmark in two different files.
